requirements.txt
requests>=2.28.1
beautifulsoup4>=4.11.1
html2text>=2020.1.16
python-dotenv>=0.20.0

# For the LLM example with free models
langchain>=0.0.267
chromadb>=0.4.13
pyyaml>=6.0
sentence-transformers>=2.2.0

# Optional: For the vector search example
numpy>=1.20.0 